cloudfunctions/past_games/requirements.txt
functions-framework==3.*
google-cloud-pubsub
google-cloud-bigquery
